Added configuration option for mipmaps (at LebedevRI's suggestion)
Look for "use embedded thumbnails if at least" in the GUI options
(embedded_thumb_minimum).  This sort of conflicts with, or subsumes,
never_use_embedded_thumb - but I am waiting on an answer for whether
or not I should absorb this option in.

In effect, this generalizes "don't use embedded preview JPEG but
half-size raw" to have options in between true and false; rather than
'never use them' or 'always use them', it lets the user pick based on
resolution.